+++ ui/litellm-dashboard/tests/top_key_view.test.tsx
@@ -1,6 +1,6 @@
 import { describe, it, expect, vi, beforeEach } from "vitest";
 import { renderWithProviders, screen, fireEvent } from "./test-utils";
-import TopKeyView from "../src/components/Usage/components/EntityUsage/top_key_view";
+import TopKeyView from "../src/components/Usage/components/EntityUsage/TopKeyView";
 import { TagUsage } from "../src/components/Usage/types";
 
 // Mock the networking module
